+++ src/app/(home)/layout.tsx
@@ -6,11 +6,11 @@ import { Avatar, Box, Container, Flex } from "@mantine/core";
 
 import { prisma } from "@/lib/prisma";
 
-import CraeteButton from "./create-button";
-import LinkButton from "./link-button";
+import CraeteButton from "./_components/create-button";
+import LinkButton from "./_components/link-button";
 
 export const metadata: Metadata = {
-  title: "Timeline",
+  title: "Timeline | All",
   description: "Generated by create next app",
 };
